+++ src/components/s3/bucketList/BucketListHeader.tsx
@@ -1,4 +1,3 @@
-import React, { useCallback, useState } from "react";
 import useModal from "@hooks/useModal";
 
 import {
@@ -8,22 +7,10 @@ import {
   GridToolbarDensitySelector,
   GridRowId,
 } from "@mui/x-data-grid";
-import {
-  AlertColor,
-  Button,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogContentText,
-  DialogTitle,
-  Link,
-  TextField,
-} from "@mui/material";
+import { AlertColor, Button } from "@mui/material";
 
 import AddCircleIcon from "@mui/icons-material/AddCircle";
-
-import apiRoutes from "@configs/apiRoutes";
-import { useRouter } from "next/router";
+import CreateBucketDialog from "./CreateBucketDialog";
 
 interface IBucketListHeaderProps {
   bucketList: any;
@@ -36,53 +23,14 @@ const BucketListHeader: React.FC<IBucketListHeaderProps> = ({
   bucketList,
   selectedIds,
   setSnackbarSeverity,
-  setSnackbarMsg: setCreationMsg,
-  openSnackbar: openCreationSnackbar,
+  setSnackbarMsg,
+  openSnackbar,
 }) => {
-  const [bucketName, setBucketName] = useState("");
   const {
     isModalOpen: isCreateBucketDialogOpen,
     openModal: openCreateBucketDialog,
     closeModal: closeCreateBucketDialog,
   } = useModal();
-
-  const router = useRouter();
-
-  const inputRef = React.useRef<HTMLInputElement>();
-  React.useEffect(() => {
-    const timeout = setTimeout(() => {
-      if (inputRef.current) inputRef.current.focus();
-    }, 100);
-
-    return () => {
-      clearTimeout(timeout);
-    };
-  }, [isCreateBucketDialogOpen]);
-
-  const createBucket = useCallback(async () => {
-    const response = await fetch(
-      `${apiRoutes.ui.s3.createBucket}?bucket=${bucketName}`
-    );
-    if (response.status !== 200) {
-      console.debug(response);
-      setSnackbarSeverity("error");
-      setCreationMsg(`Error while creating "${bucketName}"`);
-      openCreationSnackbar();
-      return;
-    }
-
-    const { data } = await response.json();
-    setSnackbarSeverity("success");
-    setCreationMsg(`"${bucketName}" successfully created`);
-    openCreationSnackbar();
-    router.replace(router.asPath);
-  }, [
-    bucketName,
-    openCreationSnackbar,
-    router,
-    setCreationMsg,
-    setSnackbarSeverity,
-  ]);
 
   return (
     <>
@@ -108,47 +56,13 @@ const BucketListHeader: React.FC<IBucketListHeaderProps> = ({
           <GridToolbarDensitySelector />
         </div>
       </GridToolbarContainer>
-      <Dialog
-        open={isCreateBucketDialogOpen}
-        onClose={closeCreateBucketDialog}
-        maxWidth="md"
-        PaperProps={{ sx: { width: "30%" } }}
-      >
-        <DialogTitle>Create Bucket</DialogTitle>
-        <DialogContent>
-          <TextField
-            inputRef={inputRef}
-            id="name"
-            label="Bucket Name"
-            autoFocus
-            margin="dense"
-            type="text"
-            fullWidth
-            variant="standard"
-            onChange={(event) => setBucketName(event.currentTarget.value)}
-            onKeyPress={(event) => {
-              if (event.key === "Enter") {
-                event.preventDefault();
-                createBucket();
-              }
-            }}
-          />
-          <DialogContentText>
-            <Link
-              href="https://docs.aws.amazon.com/AmazonS3/latest/userguide/bucketnamingrules.html"
-              target={"_blank"}
-            >
-              Naming Rules
-            </Link>
-          </DialogContentText>
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={closeCreateBucketDialog}>Cancel</Button>
-          <Button variant="contained" onClick={createBucket}>
-            Create
-          </Button>
-        </DialogActions>
-      </Dialog>
+      <CreateBucketDialog
+        isDialogOpen={isCreateBucketDialogOpen}
+        closeDialog={closeCreateBucketDialog}
+        setSnackbarSeverity={setSnackbarSeverity}
+        setCreationMsg={setSnackbarMsg}
+        openCreationSnackbar={openSnackbar}
+      />
     </>
   );
 };
